# internal/controller/transaction.go
package controller

import (
	"errors"
	"net/http"
	"strconv"
	"time"

	"hodlbook/internal/models"
	"hodlbook/internal/repo"

	"github.com/gin-gonic/gin"
	"gorm.io/gorm"
)

// ListTransactions godoc
// @Summary List transactions
// @Description Get a list of transactions with optional filters
// @Tags transactions
// @Produce json
// @Param limit query int false "Limit"
// @Param offset query int false "Offset"
// @Param asset_id query int false "Asset ID"
// @Param type query string false "Transaction type"
// @Param start_date query string false "Start date (YYYY-MM-DD)"
// @Param end_date query string false "End date (YYYY-MM-DD)"
// @Success 200 {object} repo.TransactionListResult
// @Failure 500 {object} map[string]string
// @Router /api/transactions [get]
func (c *Controller) ListTransactions(ctx *gin.Context) {
	filter := repo.TransactionFilter{}

	if limitStr := ctx.Query("limit"); limitStr != "" {
		if limit, err := strconv.Atoi(limitStr); err == nil {
			filter.Limit = limit
		}
	}
	if offsetStr := ctx.Query("offset"); offsetStr != "" {
		if offset, err := strconv.Atoi(offsetStr); err == nil {
			filter.Offset = offset
		}
	}
	if assetIDStr := ctx.Query("asset_id"); assetIDStr != "" {
		if assetID, err := strconv.ParseInt(assetIDStr, 10, 64); err == nil {
			filter.AssetID = &assetID
		}
	}
	if txType := ctx.Query("type"); txType != "" {
		filter.Type = txType
	}
	if startDateStr := ctx.Query("start_date"); startDateStr != "" {
		if startDate, err := time.Parse("2006-01-02", startDateStr); err == nil {
			filter.StartDate = &startDate
		}
	}
	if endDateStr := ctx.Query("end_date"); endDateStr != "" {
		if endDate, err := time.Parse("2006-01-02", endDateStr); err == nil {
			endDate = endDate.Add(24*time.Hour - time.Second)
			filter.EndDate = &endDate
		}
	}

	result, err := c.repo.ListTransactions(filter)
	if err != nil {
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to fetch transactions"})
		return
	}
	ctx.JSON(http.StatusOK, result)
}

// GetTransaction godoc
// @Summary Get a transaction by ID
// @Description Get a single transaction by its ID
// @Tags transactions
// @Produce json
// @Param id path int true "Transaction ID"
// @Success 200 {object} models.Transaction
// @Failure 400 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Router /api/transactions/{id} [get]
func (c *Controller) GetTransaction(ctx *gin.Context) {
	id, err := strconv.ParseInt(ctx.Param("id"), 10, 64)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Invalid transaction ID"})
		return
	}

	tx, err := c.repo.GetTransactionByID(id)
	if err != nil {
		ctx.JSON(http.StatusNotFound, gin.H{"error": "Transaction not found"})
		return
	}

	ctx.JSON(http.StatusOK, tx)
}

// CreateTransaction godoc
// @Summary Create a new transaction
// @Description Create a new transaction with the provided data
// @Tags transactions
// @Accept json
// @Produce json
// @Param transaction body models.Transaction true "Transaction data"
// @Success 201 {object} models.Transaction
// @Failure 400 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /api/transactions [post]
func (c *Controller) CreateTransaction(ctx *gin.Context) {
	var tx models.Transaction
	if err := ctx.ShouldBindJSON(&tx); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Invalid input", "details": err.Error()})
		return
	}

	if tx.Timestamp.IsZero() {
		tx.Timestamp = time.Now()
	}

	if err := c.repo.CreateTransaction(&tx); err != nil {
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create transaction"})
		return
	}

	ctx.JSON(http.StatusCreated, tx)
}

// UpdateTransaction godoc
// @Summary Update a transaction
// @Description Update an existing transaction by its ID
// @Tags transactions
// @Accept json
// @Produce json
// @Param id path int true "Transaction ID"
// @Param transaction body models.Transaction true "Transaction data"
// @Success 200 {object} models.Transaction
// @Failure 400 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /api/transactions/{id} [put]
func (c *Controller) UpdateTransaction(ctx *gin.Context) {
	id, err := strconv.ParseInt(ctx.Param("id"), 10, 64)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Invalid transaction ID"})
		return
	}

	if _, err = c.repo.GetTransactionByID(id); err != nil {
		ctx.JSON(http.StatusNotFound, gin.H{"error": "Transaction not found"})
		return
	}

	var tx models.Transaction
	if err := ctx.ShouldBindJSON(&tx); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Invalid input", "details": err.Error()})
		return
	}

	tx.ID = id
	if err := c.repo.UpdateTransaction(&tx); err != nil {
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to update transaction"})
		return
	}

	ctx.JSON(http.StatusOK, tx)
}

// DeleteTransaction godoc
// @Summary Delete a transaction
// @Description Delete a transaction by its ID
// @Tags transactions
// @Param id path int true "Transaction ID"
// @Success 204
// @Failure 400 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /api/transactions/{id} [delete]
func (c *Controller) DeleteTransaction(ctx *gin.Context) {
	id, err := strconv.ParseInt(ctx.Param("id"), 10, 64)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Invalid transaction ID"})
		return
	}

	if err := c.repo.DeleteTransaction(id); err != nil && !errors.Is(err, gorm.ErrRecordNotFound) {
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to delete transaction"})
		return
	}

	ctx.Status(http.StatusNoContent)
}